refactor: move CI types and methods in is own space
